refactor: isolate framework detection
