--- cli/src/modules/compile/types.ts
@@ -1,1 +1,0 @@
-export type Framework = "foundry" | "hardhat";
